Moved unit tests from ObjectTree::move to Directory
